+++ frontend/src/components/Card/card-svgs/index.ts
